Fixing of build warnings - Converting implicit casts to explicit ones - Renaming local variables hiding class members and local variables which have multiple declarations with the same name but different scope - Removed unused function RIFFChunk::ReadString because it requires variable length arrays (limited compiler support) - Removed unused member int SFZVoice::trigger

+++ module/SFZero/SFZero/SFZSample.cpp
@@ -18,8 +18,8 @@ bool SFZSample::load(AudioFormatManager* formatManager)
 	StringPairArray* metadata = &reader->metadataValues;
 	int numLoops = metadata->getValue("NumSampleLoops", "0").getIntValue();
 	if (numLoops > 0) {
-		loopStart = metadata->getValue("Loop0Start", "0").getLargeIntValue();
-		loopEnd = metadata->getValue("Loop0End", "0").getLargeIntValue();
+		loopStart = (unsigned long)metadata->getValue("Loop0Start", "0").getLargeIntValue();
+		loopEnd = (unsigned long)metadata->getValue("Loop0End", "0").getLargeIntValue();
 		}
 	delete reader;
 	return true;
